Refactor: shared
