+++ src/preprocessors/optimize-imports.ts
@@ -2,8 +2,8 @@ import MagicString from "magic-string";
 import type { ImportDeclaration } from "svelte/compiler";
 import { parse, walk } from "svelte/compiler";
 import type { SveltePreprocessor } from "svelte/types/compiler/preprocess";
-import { components } from "./component-index";
-import { CarbonSvelte } from "./constants";
+import { components } from "../component-index";
+import { CarbonSvelte } from "../constants";
 
 function rewriteImport(
   s: MagicString,
